refactor: merge date-helpers into date-formatting module
